pyproject.toml
[project]
name = "izip2pdf"
version = "0.1.8"
description = "convert zip file including image to pdf file"
authors = [{ name = "hy20191108", email = "zwwp9976@gmail.com" }]
dependencies = [
    "img2pdf>=0.5.1",
    "tqdm>=4.66.4",
    "natsort>=6.2.0",
    "pillow>=10.3.0",
    "pillow-avif-plugin>=1.4.3",
    "pillow-heif>=0.16.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[project.scripts]
izip2pdf = "izip2pdf.izip2pdf:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/izip2pdf"]

[project.urls]
Homepage = "https://github.com/hy20191108/izip2pdf"
